+++ app/src/main/cpp/IDemux.h
@@ -27,8 +27,8 @@ public:
 
     virtual void Close() = 0;
 
-    //seek 位置 pos 0.0~1.0
-    virtual bool Seek(double pos) = 0;
+    //seek 位置 毫秒级别的拖拽
+    virtual bool Seek(long pos) = 0;
 
     // 解封装后的一帧数据，内存由调用者清理
     virtual XData Read() = 0;
